# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Учёба с Лучиком API",
    "version": "0.1.0",
    "description": "Student mini-app API. MAX initData is exchanged for a short-lived bearer session. Schedules are synthetic, and lesson generation uses RouterAI."
  },
  "paths": {
...
    "/api/auth/max": {
      "post": {
        "summary": "Max Auth",
        "operationId": "max_auth_api_auth_max_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token"
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_api_register_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "student"
                  ],
                  "properties": {
                    "student": {}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "student",
                    "lesson",
                    "mood",
                    "daily",
                    "activity"
                  ],
                  "properties": {
                    "name": {},
                    "student": {},
                    "lesson": {},
                    "mood": {},
                    "daily": {},
                    "activity": {}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/diary": {
      "get": {
        "summary": "Diary",
        "operationId": "diary_api_diary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "mock",
                    "entries"
                  ],
                  "properties": {
                    "mock": {
                      "type": "boolean",
                      "const": true
                    },
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_history_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "topics"
                  ],
                  "properties": {
                    "topics": {}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/history/{lesson_id}/quiz-jobs": {
      "post": {
        "summary": "Start Quiz Job",
        "operationId": "start_quiz_job_api_history__lesson_id__quiz_jobs_post",
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lesson Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/quiz-jobs/{job_id}": {
      "get": {
        "summary": "Quiz Job Status",
        "operationId": "quiz_job_status_api_quiz_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
    "/api/quizzes/{quiz_id}": {
      "get": {
        "summary": "Get Quiz",
        "operationId": "get_quiz_api_quizzes__quiz_id__get",
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quiz Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "StudentSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Login": {
        "properties": {
          "init_data": {
            "type": "string",
            "maxLength": 16000,
            "minLength": 1,
            "title": "Init Data"
          }
        },
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "Login"
      },
...
    },
    "securitySchemes": {
      "StudentSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session returned by POST /api/auth/max; valid for 24 hours."
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "luchik_admin",
        "description": "Separate admin session; only needed for schedule administration."
      },
      "MaxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "https://85-159-231-99.sslip.io",
      "description": "Public hackathon deployment"
    }
  ]
}